updated: refactored source code
