+++ Makefile
@@ -48,9 +48,9 @@ raw:
 	# build riot
 	@ mkdir -p $(DIST)
 	# Default builds UMD
-	@ $(ROLLUP) src/riot.js --format umd --config rollup.config.js > $(DIST)riot.js
-	@ $(ROLLUP) src/riot+compiler.js --format umd --config rollup.config.js > $(DIST)riot+compiler.js
-	@ $(ROLLUP) src/riot.js --format esm --config rollup.config.js > $(DIST)riot.esm.js
+	@ $(ROLLUP) src/riot.js --format umd --config rollup.config.js --file $(DIST)riot.js
+	@ $(ROLLUP) src/riot+compiler.js --format umd --config rollup.config.js --file $(DIST)riot+compiler.js
+	@ $(ROLLUP) src/riot.js --format esm --preserveModules --preserveModulesRoot src --config rollup.config.js --dir $(DIST)esm
 
 clean:
 	# clean $(DIST)
