+++ package.json
@@ -44,15 +44,14 @@
     "curri": "^1.0.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.18.6",
-    "@babel/preset-env": "^7.18.6",
+    "@babel/core": "^7.18.9",
+    "@babel/preset-env": "^7.18.9",
     "@riotjs/babel-preset": "1.0.0",
     "@rollup/plugin-commonjs": "^22.0.1",
-    "@rollup/plugin-json": "^4.1.0",
     "@rollup/plugin-node-resolve": "^13.3.0",
     "babel-plugin-istanbul": "^6.1.1",
     "chai": "^4.3.6",
-    "core-js": "^3.23.4",
+    "core-js": "^3.23.5",
     "eslint": "^8.20.0",
     "eslint-config-riot": "^3.0.0",
     "jsdom": "20.0.0",
@@ -73,16 +72,15 @@
     "typescript": "^4.7.4"
   },
   "files": [
-    "src",
+    "esm",
     "riot.js",
     "riot.d.ts",
-    "riot.esm.js",
     "riot.min.js",
     "riot+compiler.js",
     "riot+compiler.min.js"
   ],
   "main": "riot.js",
-  "module": "riot.esm.js",
-  "jsnext:main": "riot.esm.js",
+  "module": "esm/riot.js",
+  "jsnext:main": "esm/riot.js",
   "types": "./riot.d.ts"
 }
