+++ src/api/uninstall.js
@@ -1,0 +1,15 @@
+import {PLUGINS_SET} from '@riotjs/util/constants'
+import {panic} from '@riotjs/util/misc'
+
+/**
+ * Uninstall a riot plugin
+ * @param   {Function} plugin - plugin previously installed
+ * @returns {Set} the set containing all the plugins installed
+ */
+export function uninstall(plugin) {
+  if (!PLUGINS_SET.has(plugin)) panic('This plugin was never installed')
+
+  PLUGINS_SET.delete(plugin)
+
+  return PLUGINS_SET
+}
