+++ src/api/unregister.js
@@ -1,0 +1,17 @@
+import {COMPONENTS_IMPLEMENTATION_MAP} from '@riotjs/util/constants'
+import cssManager from '../core/css-manager'
+import {panic} from '@riotjs/util/misc'
+
+/**
+ * Unregister a riot web component
+ * @param   {string} name - component name
+ * @returns {Map} map containing all the components implementations
+ */
+export function unregister(name) {
+  if (!COMPONENTS_IMPLEMENTATION_MAP.has(name)) panic(`The component "${name}" was never registered`)
+
+  COMPONENTS_IMPLEMENTATION_MAP.delete(name)
+  cssManager.remove(name)
+
+  return COMPONENTS_IMPLEMENTATION_MAP
+}
