+++ src/api/version.js
@@ -1,0 +1,2 @@
+/** @type {string} current riot version */
+export const version = 'WIP'
