+++ src/api/with-types.js
@@ -1,0 +1,7 @@
+/**
+ * no-op function needed to add the proper types to your component via typescript
+ * @param {Function|Object} component - component default export
+ * @returns {Function|Object} returns exactly what it has received
+ */
+/* istanbul ignore next */
+export const withTypes = component => component
