+++ src/compiler/global-registry.js
@@ -1,0 +1,1 @@
+export const GLOBAL_REGISTRY = '__riot_registry__'
