+++ src/compiler/transpile.js
@@ -1,0 +1,4 @@
+// cheap module transpilation
+export function transpile(code) {
+  return `(function (global){${code}})(this)`.replace('export default', 'return')
+}
