+++ src/core/component-core-helpers.js
@@ -1,0 +1,7 @@
+import $ from 'bianco.query'
+
+export const COMPONENT_CORE_HELPERS = Object.freeze({
+  // component helpers
+  $(selector){ return $(selector, this.root)[0] },
+  $$(selector){ return $(selector, this.root) }
+})
